+++ CHANGELOG.md
@@ -2,6 +2,77 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [43.0.0](https://github.com/kufu/smarthr-ui/compare/v42.0.0...v43.0.0) (2024-02-28)
+
+
+### ⚠ BREAKING CHANGES
+
+* BackgroundJobsPanel を削除 (#4359)
+
+### Features
+
+* Icon に FaPaintBrush を追加 ([#4345](https://github.com/kufu/smarthr-ui/issues/4345)) ([744a012](https://github.com/kufu/smarthr-ui/commit/744a012ba102dd85d71d0780f972bd29f351ca31))
+
+
+### Bug Fixes
+
+* FormControl コンポーネントの props が DOM に表出しないようにする ([#4344](https://github.com/kufu/smarthr-ui/issues/4344)) ([0626715](https://github.com/kufu/smarthr-ui/commit/062671517fa467cac12bb330fa5c3dda57eb0b0c))
+* HeaderLink に className を渡せるように修正 ([715c315](https://github.com/kufu/smarthr-ui/commit/715c315c34667d74fceb4bbe13e219125429bee2))
+* HeaderLink に className を渡せるように修正 ([#4358](https://github.com/kufu/smarthr-ui/issues/4358)) ([a016836](https://github.com/kufu/smarthr-ui/commit/a016836043d25e805b9361f3c7f1fd7c76331bde))
+* Paginationコンポーネントの番号ボタンに不必要なlist-styleが表示されてしまう問題を修正 ([#4349](https://github.com/kufu/smarthr-ui/issues/4349)) ([f82ba44](https://github.com/kufu/smarthr-ui/commit/f82ba44005ef8e58d9db2dfc514903f476cdc093))
+* TabItem の余白を調整 ([#4364](https://github.com/kufu/smarthr-ui/issues/4364)) ([08a7651](https://github.com/kufu/smarthr-ui/commit/08a7651d1ee6e5574d2f2a8b04ca76cf5c8a780f))
+* Textareaでwidthを指定しない場合、style属性で"width=auto;"が指定されないようにする ([#4350](https://github.com/kufu/smarthr-ui/issues/4350)) ([6039bf3](https://github.com/kufu/smarthr-ui/commit/6039bf3713e3ce0ca85cc6b72ad3719fc59aac58))
+* 強制カラーモードの時、borderのスタイルが適用されるようにした ([#4365](https://github.com/kufu/smarthr-ui/issues/4365)) ([5c74edc](https://github.com/kufu/smarthr-ui/commit/5c74edc0b60cc812e108d25038f12fd1859a559c))
+
+
+* BackgroundJobsPanel を削除 ([#4359](https://github.com/kufu/smarthr-ui/issues/4359)) ([b9a0b64](https://github.com/kufu/smarthr-ui/commit/b9a0b64ddbd3289bbbeae1d414622506326fa28e))
+
+## [42.0.0](https://github.com/kufu/smarthr-ui/compare/v41.3.0...v42.0.0) (2024-02-20)
+
+
+### ⚠ BREAKING CHANGES
+
+* RightFixedNote と DialogBase を削除 (#4318)
+
+### Bug Fixes
+
+* DOM に themes="[object Object]" がつかないように修正 ([#4293](https://github.com/kufu/smarthr-ui/issues/4293)) ([76df903](https://github.com/kufu/smarthr-ui/commit/76df90367b25bfe74486203537a6b442231f9cc8))
+* PageCounterの-がスクリーンリーダーで読み上げられない問題を修正 ([#4297](https://github.com/kufu/smarthr-ui/issues/4297)) ([a4e4ed6](https://github.com/kufu/smarthr-ui/commit/a4e4ed631b434c55d161120361e6ecf25f7c8dee))
+* TableReel内の要素のサイズが動的に変化したときにもfixedの計算をするようにした ([#4316](https://github.com/kufu/smarthr-ui/issues/4316)) ([0b0b6de](https://github.com/kufu/smarthr-ui/commit/0b0b6ded515fa76c10e0c194f6b83b382807fbec))
+
+
+* RightFixedNote と DialogBase を削除 ([#4318](https://github.com/kufu/smarthr-ui/issues/4318)) ([d3e4ab3](https://github.com/kufu/smarthr-ui/commit/d3e4ab395935af5317c8a41b776812a74a5c391d))
+
+## [41.3.0](https://github.com/kufu/smarthr-ui/compare/v41.2.0...v41.3.0) (2024-02-14)
+
+
+### Features
+
+* preset に border の簡略表記を追加し、overflow の初期化を削除 ([#4298](https://github.com/kufu/smarthr-ui/issues/4298)) ([3b59865](https://github.com/kufu/smarthr-ui/commit/3b598657ee91b9ab8edce7660b1777f041b125d5))
+* RightFixedNoteにloadingとsubmittingの状態を追加 ([#4306](https://github.com/kufu/smarthr-ui/issues/4306)) ([4fff8eb](https://github.com/kufu/smarthr-ui/commit/4fff8ebfbfa23137529e0a61db60549301932d19))
+
+
+### Bug Fixes
+
+* broken character in ComboBox story ([#4309](https://github.com/kufu/smarthr-ui/issues/4309)) ([72e8516](https://github.com/kufu/smarthr-ui/commit/72e8516dfbac116d5cc5ddef1f23fd16e9409a38))
+* FilterDropdown[disabled] 時のアイコン色を修正 ([#4299](https://github.com/kufu/smarthr-ui/issues/4299)) ([758c1db](https://github.com/kufu/smarthr-ui/commit/758c1db81d8fa29bfc43a064c5dacba410287d8a))
+* FormControl, Fieldset の errorMessages に空配列が渡された場合、不必要な要素をレンダリングしないように修正 ([#4288](https://github.com/kufu/smarthr-ui/issues/4288)) ([79ff6ab](https://github.com/kufu/smarthr-ui/commit/79ff6abbc974bf952ef77581e0d5e8c9b327263d))
+* ハイコントラストモード時、AppNaviのAnchorに下線を表示する ([#4292](https://github.com/kufu/smarthr-ui/issues/4292)) ([25c984c](https://github.com/kufu/smarthr-ui/commit/25c984c1ec25e00cd4cad704fa6c6432313143ad))
+
+## [41.2.0](https://github.com/kufu/smarthr-ui/compare/v41.1.2...v41.2.0) (2024-02-06)
+
+
+### Features
+
+* Fieldset のタイトルを視覚的に隠す dangerouslyTitleHidden を追加 ([#4253](https://github.com/kufu/smarthr-ui/issues/4253)) ([98dc367](https://github.com/kufu/smarthr-ui/commit/98dc367812fc2a136525dd4c708b936679f4790d))
+* loaderの色をmainに変更。差異がなくなるのでprefres-contrast: moreでの指定は削除。 ([#4251](https://github.com/kufu/smarthr-ui/issues/4251)) ([77a6a6d](https://github.com/kufu/smarthr-ui/commit/77a6a6d97766c4f60cce85d0258408940f82e042))
+* SelectのValueに英字などが来た場合、ディセンダー部分が見切れるのを修正 ([#4261](https://github.com/kufu/smarthr-ui/issues/4261)) ([dccb116](https://github.com/kufu/smarthr-ui/commit/dccb1165ccfe3bc462a486bfe7cca70f12ea0618))
+
+
+### Bug Fixes
+
+* AccordionPanel の max-height に余計なスタイルがあたっていたため削除 ([#4250](https://github.com/kufu/smarthr-ui/issues/4250)) ([de478c9](https://github.com/kufu/smarthr-ui/commit/de478c96df474721879d84a2a24b787863d6d496))
+
 ### [41.1.2](https://github.com/kufu/smarthr-ui/compare/v41.1.1...v41.1.2) (2024-01-29)